Broke out Artifacts specific to Huffman encoding/decoding into their own folder
***
First commit after cherry-picking the commit history of
non_core_functional_refactor branch to the end of main due
to issues between WSL and git repo. non_core_function_refactor
was initially a branch of master which was effectively a clone
of the most recent version of main at the time.
***